functions/requirements.txt
functions-framework==3.5.0
psycopg2-binary==2.9.9
google-cloud-scheduler==3.14.1
google-auth==2.25.2
google-api-python-client==1.12.5
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
